some changes in data model due to new requirements

+++ ModiSmile.DataAccess/Repositories/IEventRepository.cs
@@ -5,7 +5,7 @@ namespace ModiSmile.DataAccess.Repositories
 {
     public interface IEventRepository
     {
-        long? GetUserEvents(int? aggregateId, string aggregateType, string[] userIds,string clientId, ActionTypes action, DateTime? from, DateTime? to);
+        double? GetUserEvents(int? aggregateId, string aggregateType, string[] userIds,string clientId, ActionTypes action, DateTime? from, DateTime? to);
         void Insert(Event newEvent);
     }
 }
